Day6Part1: Did some additional experimentation with manually drawing lines between each point;
in ascending order distance and without overlaps.

The result is effectively triangulating a set of points
(i.e, creating a maximal planar graph).

It looks as if we could implement an algorithm to, given a point, calculate the area of all neighboring triangles. 

This may at least provide a indicator of relative size to at least narrow down the point with the largest owned area. Although preferably we could calculate the area directly based on half-way points to all directly connected neighbors.

After performing further research, it seems that we're dealing with a Voronoi diagram, which is related to Delaunay triangulation. Looks like I'm on the right track!